browser: make the CSS cascade match a browser, verified against Chromium
The cascade was loadable but not faithful, and the failures were silent. This
adds a way to see and check them, then fixes what it found.

tools/cssdiff compares this package's computed styles with a real Chromium,
element by element, on the same page: Chromium answers over CDP, the Go browser
answers through `gobrowser get --eval`, and the two are matched by DOM position.
It is a separate module so chromedp never reaches the browser package. With it,
one page at a time, the following were found and fixed:

- var() and CSS custom properties did not exist, so every declaration that used
  them was dropped. This was the big one: Wikipedia gets its text colour from
  "color: var(--color-base, #202122)" on body and Bootstrap 5 themes every
  component with --bs-*, so such pages fell back to user-agent defaults. Now
  custom properties cascade, inherit and substitute, including fallbacks and
  nested references (89.5% -> 0.0% of Wikipedia elements differing in colour).

- @media was evaluated by splitting on " and ", which never matched a query
  written across lines ("only screen\nand (min-width: 300px)\nand ..."), and an
  unparsed query was treated as matching. Both applied every mobile rule on a
  desktop page: Hacker News rendered 9pt text with block pagelinks. Queries are
  now tokenised (media types, not/only, and/or, width and height families,
  orientation, device-* aliases, the (width >= 600px) range form), and an
  unparseable query does not match.

- font-size resolved percentages and em against the element's own user-agent
  size instead of its parent's, so "h1 { font-size: 2em }" was 64px not 32px and
  Bootstrap's "small { font-size: 87% }" was 87% of 13px not of 16px. Absolute
  keywords now come from the medium-keyed table (large is 18px at any parent
  size, small is 13px) and the user-agent <small> is "smaller".

- The 13px monospace size applied even when an author font-family replaced it.
  A browser keeps it only while the computed family is exactly the monospace
  keyword, which is why Wikipedia's
  "pre,code,tt,kbd,samp{font-family:monospace,monospace}" gives 16px code.

- The default text colour was #202124 rather than black, <th> was not bold,
  <center> was inline, tables had display:block, form controls were hidden, and
  a floated element was not blockified in its computed display.

- HTML presentational attributes (bgcolor, color, align) were ignored, so old
  table layouts had no colour; a document without a doctype is now styled in
  quirks mode, where tables take the medium font size.

- document.write appended to the body instead of inserting at the writing
  script's position, so content written mid-document (quotes.toscrape.com's
  quote list) landed after the footer. Chromium confirmed the correct order.

- getComputedStyle resolved media queries at a hardcoded 1280px; it now uses the
  width the page is laid out at.

- Style loading failures and dropped selectors were invisible. --debug now logs
  each stylesheet with its size, each failure, the rules per sheet and samples
  of the selectors cascadia could not compile.

- The NodeFilter constants were ints, which does not build for GOARCH=386 or
  arm because SHOW_ALL is 0xFFFFFFFF.

Measured with tools/cssdiff, on every element both engines produce, across seven
properties (font-size, colour, background, weight, style, display, alignment):
news.ycombinator.com 817/817 agree, quotes.toscrape.com/js 109/109 agree,
Wikipedia's Go article 6703 elements with font-size 29.6% -> 0.3% differing and
colour 89.5% -> 0.0%. Remaining Wikipedia gaps are :is()/:has() selectors and
flex/grid layout. Loading a page and extracting text still never fetch a
stylesheet, and the browser builds for all nine cross-compilation targets.

+++ browser/jsweb.go
@@ -631,7 +631,9 @@ const (
 // newNodeFilter exposes the NodeFilter constants used with TreeWalker.
 func (e *jsEnv) newNodeFilter() *goja.Object {
 	o := e.vm.NewObject()
-	for k, v := range map[string]int{
+	// float64, because these are JavaScript numbers and SHOW_ALL does not fit
+	// in a 32-bit int.
+	for k, v := range map[string]float64{
 		"FILTER_ACCEPT": 1,
 		"FILTER_REJECT": 2,
 		"FILTER_SKIP":   3,
